+++ libgfx/include/detail/triangle.hh
@@ -23,7 +23,7 @@ class TriangleRenderer : public IBatchRenderer {
 
 public:
     explicit TriangleRenderer(gfx::Window& window);
-    void draw(int x0, int y0, int x1, int y1, int x2, int y2, gfx::Color color, glm::mat4 view);
+    void draw(float x0, float y0, float x1, float y1, float x2, float y2, gfx::Color color, glm::mat4 view);
     void flush() override;
 
 };
